facts/requirements.txt
langchain==0.0.353
python-dotenv==0.10.3
openai==0.28.1
tiktoken